Moves handler out to memory file
