+++ torchtune/training/memory.py
@@ -7,7 +7,7 @@
 import gc
 import logging
 
-from typing import Any, Callable, Dict, Set, Type, Union
+from typing import Any, Callable, Dict, List, Set, Type, Union
 
 import torch
 from torch import nn
@@ -304,3 +304,68 @@ def log_memory_stats(
         f"\n\t{device_support.device_name} peak memory reserved: {stats['peak_memory_reserved']:.2f} GiB"
         f"\n\t{device_support.device_name} peak memory active: {stats['peak_memory_active']:.2f} GiB"
     )
+
+
+class Float8Handler:
+    def __init__(self, enable_fp8_training: bool, dp_shard: int):
+        self.enabled = enable_fp8_training
+
+        if not self.enabled:
+            return
+        try:
+            from torchao.float8 import Float8LinearConfig
+            from torchao.float8.config import CastConfig, ScalingType
+        except ImportError as e:
+            raise ImportError(
+                "torchao is not installed. Please install it to use float8 linear layers."
+            ) from e
+
+        dp_shard_enabled = dp_shard > 1
+        # Mutates the model inplace replacing instances of torch.nn.Linear with Float8Linear
+
+        # in torchtitan they also check if this is explicitly set - this implementation is
+        # more opinionated if fsdp
+        enable_fsdp_float8_all_gather = dp_shard_enabled
+        self.precompute_scale = enable_fsdp_float8_all_gather
+
+        self.config = Float8LinearConfig(
+            enable_fsdp_float8_all_gather=enable_fsdp_float8_all_gather,
+            force_recompute_fp8_weight_in_bwd=True,
+            cast_config_input=CastConfig(scaling_type=ScalingType.DYNAMIC),
+            cast_config_weight=CastConfig(scaling_type=ScalingType.DYNAMIC),
+            cast_config_grad_output=CastConfig(scaling_type=ScalingType.DYNAMIC),
+        )
+        self.enabled = True
+
+    def convert_to_float8_training(self, model: nn.Module):
+        """
+        This function converts the linear layers of `model` to `Float8Linear`.
+        Note that today, only dynamic tensor scaling (the default) is supported.
+        This will mutate the model inplace.
+        """
+        if not self.enabled:
+            return
+
+        from torchao.float8 import convert_to_float8_training
+
+        # Mutates the model inplace replacing instances of nn.Linear with Float8Linear
+        convert_to_float8_training(
+            model,
+            config=self.config,
+            module_filter_fn=lambda mod, fqn: fqn != "output",
+        )
+
+    def precompute_float8_dynamic_scale_for_fsdp(
+        self, model: Union[nn.Module, List[nn.Module]]
+    ):
+        if not self.enabled:
+            return
+
+        if not self.precompute_scale:
+            return
+
+        from torchao.float8 import precompute_float8_dynamic_scale_for_fsdp
+
+        models = [model] if isinstance(model, nn.Module) else model
+        for m in models:
+            precompute_float8_dynamic_scale_for_fsdp(m)
